+++ include/subgraph.h
@@ -13,9 +13,9 @@ enum STATUS
 
 struct snode
 {
-  float   path_val;             /* path value                                           */
-  float   dens;                 /* node density                                         */
-  float   radius;               /* maximum distance among the k-nearest neighbors in
+  double   path_val;             /* path value                                           */
+  double   dens;                 /* node density                                         */
+  double   radius;               /* maximum distance among the k-nearest neighbors in
                                  * the training set. it is used to propagate
                                  * clustering labels to testing nodes)                  */
   int     label;                /* node label                                           */
@@ -23,7 +23,7 @@ struct snode
   int     pred;                 /* predecessor node                                     */
   int     label_true;           /* true label if it is known                            */
   int     position;             /* index in the feature space                           */
-  float  *feat;                 /* feature vector                                       */
+  double  *feat;                 /* feature vector                                       */
   enum STATUS  status;          /* 0 - nothing, 1 - prototype                           */
 
   int     nplatadj;             /* holds the amount of adjacent nodes on plateaus
@@ -63,19 +63,19 @@ struct subgraph
   int     feat_n;                /* number of features                                  */
   int     k_best;                /* number of adjacent nodes                            */
   int     label_n;               /* number of clusters                                  */
-  float   df;                    /* radius in the feature space for density computation */
-  float   dens_min;              /* minimum density value                               */
-  float   dens_max;              /* maximum density value                               */
-  float   k;                     /* constant for opf_pdf computation                    */
+  double   df;                    /* radius in the feature space for density computation */
+  double   dens_min;              /* minimum density value                               */
+  double   dens_max;              /* maximum density value                               */
+  double   k;                     /* constant for opf_pdf computation                    */
   int    *ordered_list_of_nodes; /* store the list of nodes in the increasing order
                                   * of cost for speeding up supervised classification.
                                   */
 
-  float *feat_data; /* memory pointer to all features data */
-  float (*arc_weight) (float *f1, float *f2, int n);
+  double *feat_data; /* memory pointer to all features data */
+  double (*arc_weight) (double *f1, double *f2, int n);
 
   /* precomputed distance matrix */
-  float *pdist;
+  double *pdist;
   int pdist_train_stride; /* stride of training (pdist) distance matrix */
 };
 
@@ -83,18 +83,18 @@ struct subgraph
 
 struct subgraph * subgraph_create       (int node_n);     /* allocates nodes without features        */
 void              subgraph_destroy      (struct subgraph ** sg); /* deallocates memory for subgraph         */
-int               subgraph_set_feature  (struct subgraph *sg, float *feat, int *label, int feat_n);
-void              subgraph_set_metric   (struct subgraph *sg, float (*arc_weight) (float *f1, float *f2, int n), enum METRIC m);
+int               subgraph_set_feature  (struct subgraph *sg, double *feat, int *label, int feat_n);
+void              subgraph_set_metric   (struct subgraph *sg, double (*arc_weight) (double *f1, double *f2, int n), enum METRIC m);
 
 int               subgraph_set_precomputed_distance
-                                        (struct subgraph *sg, float *dist, int *label);
+                                        (struct subgraph *sg, double *dist, int *label);
 
 void              subgraph_pdf_evaluate (struct subgraph * sg);
 
 void              subgraph_resize       (struct subgraph * sg, int node_n);
 
 inline static
-float subgraph_get_distance (struct subgraph * sg, struct snode *i, struct snode *j)
+double subgraph_get_distance (struct subgraph * sg, struct snode *i, struct snode *j)
 {
   if (sg->pdist)
     return PDISTANCE(sg, i->position, j->position);
