+++ include/supervised.h
@@ -7,12 +7,12 @@
 /* training function */
 void opf_supervised_train (struct subgraph * sg);
 
-void opf_supervised_train_iterative (struct subgraph *sg, float split);
+void opf_supervised_train_iterative (struct subgraph *sg, double split);
 
-void opf_supervised_train_agglomerative (struct subgraph *sg, float split);
+void opf_supervised_train_agglomerative (struct subgraph *sg, double split);
 
 /* classification function: it simply
  * classifies samples from sg */
-void opf_supervised_classify (struct subgraph * sgtrain, float *feat, int sample_n, int *label);
+void opf_supervised_classify (struct subgraph * sgtrain, double *feat, int sample_n, int *label);
 
 #endif
